First step in porting to Nexys 3
Changed the port mappings in the UCF to match the Nexys 3 UCF. Generated
a new ~21 MHz clock core that runs on Nexys 3. I can program the FPGA
and the loader application loads a ROM without complaining and seems to
send keys over UART, however the displayshows either nothing or a white
line down the right side and across the bottom of the screen, depending
on the switches on the board.
NB. Quite a lot of pins have been left unconnected due to hardware
differences between Nexys 3 and 4.